wip: finish fixup validity checks
