+++ webui/src/Components/NumberInputField.tsx
@@ -1,9 +1,8 @@
-import React, { useEffect, useCallback, useState } from 'react'
+import React, { useCallback, useState } from 'react'
 import { CCol, CFormInput, CFormRange, CRow } from '@coreui/react'
 
 interface NumberInputFieldProps {
 	label?: React.ReactNode
-	required?: boolean
 	min?: number
 	max?: number
 	step?: number
@@ -11,13 +10,12 @@ interface NumberInputFieldProps {
 	range?: boolean
 	value: number
 	setValue: (value: number) => void
-	setValid?: (valid: boolean) => void
 	disabled?: boolean
+	checkValid?: (value: number) => boolean
 }
 
 export function NumberInputField({
 	label,
-	required,
 	min,
 	max,
 	step,
@@ -25,44 +23,10 @@ export function NumberInputField({
 	range,
 	value,
 	setValue,
-	setValid,
 	disabled,
+	checkValid,
 }: NumberInputFieldProps) {
 	const [tmpValue, setTmpValue] = useState<string | number | null>(null)
-
-	// Check if the value is valid
-	const isValueValid = useCallback(
-		(val: string | number) => {
-			if (val === '') {
-				// If required, it must not be empty
-				if (required) {
-					return false
-				}
-			} else {
-				const valNum = Number(val)
-				// If has a value, it must be a number
-				if (isNaN(valNum)) {
-					return false
-				}
-
-				// Verify the value range
-				if (min !== undefined && valNum < min) {
-					return false
-				}
-				if (max !== undefined && valNum > max) {
-					return false
-				}
-			}
-
-			return true
-		},
-		[required, min, max]
-	)
-
-	// If the value is undefined, populate with the default. Also inform the parent about the validity
-	useEffect(() => {
-		setValid?.(isValueValid(value))
-	}, [isValueValid, value, setValid])
 
 	const onChange = useCallback(
 		(e: React.FormEvent<HTMLInputElement>) => {
@@ -70,9 +34,8 @@ export function NumberInputField({
 			const processedValue = isNaN(parsedValue) ? e.currentTarget.value : parsedValue
 			setTmpValue(processedValue)
 			setValue(Number(processedValue))
-			setValid?.(isValueValid(processedValue))
 		},
-		[setValue, setValid, isValueValid]
+		[setValue]
 	)
 
 	// Render the input
@@ -85,7 +48,7 @@ export function NumberInputField({
 			min={min}
 			max={max}
 			step={step ?? 'any'}
-			style={{ color: !isValueValid(tmpValue ?? value) ? 'red' : undefined }}
+			style={{ color: !!checkValid && !checkValid(Number(tmpValue ?? value)) ? 'red' : undefined }}
 			title={tooltip}
 			onChange={onChange}
 			onFocus={() => setTmpValue(value ?? '')}
